move AWS examples to project directory and update namespace
