+++ src/Aws/examples/AwsClientApp/composer.json
