+++ src/Aws/examples/collector/config.yaml
