+++ src/Aws/examples/sampleApp2/Service2.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * Copyright The OpenTelemetry Authors
  *
@@ -21,17 +22,17 @@ namespace Examples\Aws\SampleApp2;
 
 require __DIR__ . '/../../../vendor/autoload.php';
 
-use OpenTelemetry\Aws\Xray\IdGenerator;
-use OpenTelemetry\Aws\Xray\Propagator;
+use OpenTelemetry\Contrib\Aws\Xray\IdGenerator;
+use OpenTelemetry\Contrib\Aws\Xray\Propagator;
 use OpenTelemetry\Contrib\OtlpGrpc\Exporter as OTLPExporter;
 use OpenTelemetry\SDK\Trace\PropagationMap;
 use OpenTelemetry\SDK\Trace\SpanProcessor\SimpleSpanProcessor;
 use OpenTelemetry\SDK\Trace\TracerProvider;
 use OpenTelemetry\Trace as API;
 
-class Service1
+class Service2
 {
-    private const LIMIT = 100;
+    private const LIMIT = 50;
 
     private $tracer;
     private $carrier;
@@ -45,19 +46,24 @@ class Service1
     {
         $Exporter = new OTLPExporter();
         $map = new PropagationMap();
-        
+
         // Create a tracer object that uses the AWS X-Ray ID Generator to
         // generate trace IDs in the correct format
         $tracer = (new TracerProvider(null, null, new IdGenerator()))
             ->addSpanProcessor(new SimpleSpanProcessor($Exporter))
             ->getTracer('io.opentelemetry.contrib.php');
-        
+
         // Extract the SpanContext from the carrier
         $spanContext = Propagator::extract($this->carrier, $map);
 
         // Do some kind of operation
         $i = 0;
+        $j = 0;
         while ($i < self::LIMIT) {
+            while ($j < self::LIMIT) {
+                $j += $i + $j;
+                $j++;
+            }
             $i++;
         }
 
@@ -66,7 +72,7 @@ class Service1
 
         // Set some dummy attributes
         $childSpan->setAttribute('service_2', 'microservice')
-                ->setAttribute('action_item', (string) $i);
+                ->setAttribute('action_item', (string) $j);
 
         // End child span
         $childSpan->end();
